connected to APIs
First I connected to the Trefle.io API, only to discover several of the fields I wanted to use had null values. Then I switched to Perenual. I've be working with the structure of the data, starting with a basic array of strings (names of sample vegetables) that I later map over to get more details for from the API. I am using Redux and had problems with the store; when one component failed due to issues with the GET requests another component would fail too.